Rendering is a bit of a conundrum.
No real way to order entities in an ECS, I suppose.
So we either add a component for each layer, which would
honestly work pretty well I bet, or we do something
brute-force like copy the data we need out, sort it,
and then draw it.  Prooooobably should do the former.